feat:add main layout (#7)
## Implement Main Layout and Header Components for Mumble Clone

### Overview
This PR introduces the Main Layout and Header components for our
Next.js-based Mumble clone. These components are crucial for
establishing the basic structure and navigation of our application,
setting the stage for further development of features and UI elements.

### Changes Included
- **Main Layout Component**: Serves as the application's structural
backbone, wrapping around page content and embedding the Header
component. It ensures a consistent layout across different pages.
- **Header Component**: Provides the primary navigation bar, including
links to home, explore, notifications, and user settings. It's designed
to be responsive and adaptable to different screen sizes. Currently Logo
and Login-Button is the only content.

### Technical Details
- Integrated clada-storybook for styling and components (currently just
the Logo component is used)
- There might be more setup necessary as the components from the design
system are not suggested for auto-import with vs-code intellisense.

### Additional Notes (not directly related to the PR)
- The Logo component seems to be too large. We might need to address
this in the design system.
- Zitadel-Config has the return URL port 3100 but nextjs defaults to
3000.
- I have been unable to login in the dev environment with (event when
the port is changed to 3100)